+++ go-functions/tem_lab_xxns/tem_lab_xxns.go
@@ -36,7 +36,7 @@ func Parse(w http.ResponseWriter, r *http.Request) {
 
 type temLabxxnsStruct struct {
 	ID           uint8   `json:"id"`
-	BatteryLevel float32 `json:"battery_level"`
+	BatteryLevel uint8   `json:"battery_level"`
 	InternalData string  `json:"internal_data"`
 	Temperature  float32 `json:"temperature"`
 }
@@ -45,7 +45,7 @@ func (t *temLabxxnsStruct) parse(payload []byte) {
 	length := len(payload)
 	t.ID = uint8(payload[0])
 	// battery level expressed in 1/254 %
-	t.BatteryLevel = float32(payload[1]) / 254.0 * 100
+	t.BatteryLevel = uint8(float32(payload[1]) / 254.0 * 100.0)
 	for _, b := range payload[2 : length-2] {
 		t.InternalData += fmt.Sprintf("%02x", b)
 	}
